Split the three oversized frontend files at family boundaries and group the checked-call argument state
