+++ crates/almide-frontend/src/canonicalize/registration.rs
@@ -329,265 +329,6 @@ pub fn register_protocol_decl(env: &mut TypeEnv, name: &str, generics: &Option<V
     });
 }
 /// Protocols whose auto-derive RECURSES INTO EACH FIELD'S TYPE: deriving them on a struct/variant emits per-field work that requires the field type to ALSO satisfy the protocol. `Codec` calls `Field.encode` / `Field.decode`; `Ord`/`Hash` lower to a Rust `#[derive(Ord/Hash)]` that needs the field's Rust type to impl it. `Eq`/`Repr` are excluded — every generated struct gets `PartialEq` + a repr path unconditionally, so a field need not declare them (gating those would be a false positive).
-const FIELD_RECURSIVE_PROTOCOLS: &[&str] = &["Codec", "Ord", "Hash"];
-/// The field-type slots a structural type exposes to its derive: record fields, and every variant case's payload (tuple positions / record fields).
-fn type_field_slots(ty: &Ty) -> Vec<(String, Ty)> {
-    match ty {
-        Ty::Record { fields } | Ty::OpenRecord { fields } =>
-            fields.iter().map(|(n, t)| (n.to_string(), t.clone())).collect(),
-        Ty::Variant { cases, .. } => {
-            let mut out = Vec::new();
-            for c in cases {
-                match &c.payload {
-                    VariantPayload::Unit => {}
-                    VariantPayload::Tuple(ts) => for (i, t) in ts.iter().enumerate() {
-                        out.push((format!("{}.{}", c.name, i), t.clone()));
-                    },
-                    VariantPayload::Record(fs) => for (n, t) in fs {
-                        out.push((n.to_string(), t.clone()));
-                    },
-                }
-            }
-            out
-        }
-        _ => Vec::new(),
-    }
-}
-/// The nominal leaf types a derive must recurse into for one field type, descending through the standard containers (List/Option/Set/Map/Result via `Applied`, tuples, nested anon records). A `List[Pigment]` field under a `: Codec` type requires `Pigment` to be Codec, so the leaf is `Pigment`.
-fn collect_leaf_nominals(ty: &Ty, out: &mut Vec<Sym>) {
-    match ty {
-        Ty::Named(n, args) => {
-            out.push(*n);
-            for a in args { collect_leaf_nominals(a, out); }
-        }
-        Ty::Variant { name, .. } => out.push(*name),
-        Ty::Applied(_, args) => for a in args { collect_leaf_nominals(a, out); },
-        Ty::Tuple(elems) => for e in elems { collect_leaf_nominals(e, out); },
-        Ty::Record { fields } | Ty::OpenRecord { fields } =>
-            for (_, t) in fields { collect_leaf_nominals(t, out); },
-        _ => {}
-    }
-}
-/// Does user type `leaf` satisfy protocol `proto`? Keyed leniently: `type_protocols` is interned bare, but a cross-module field type may carry a qualified `mod.Type` name — accept either spelling. For `Codec`, a hand-written `Type.encode`/`Type.decode` pair (without a `: Codec` declaration) also satisfies the requirement, since the derive only needs those functions to exist.
-fn leaf_satisfies(env: &TypeEnv, leaf: Sym, proto: &str) -> bool {
-    let bare = leaf.as_str().rsplit('.').next().unwrap_or(leaf.as_str());
-    let declares = |name: &str| env.type_protocols.get(&sym(name))
-        .map_or(false, |s| s.contains(&sym(proto)));
-    if declares(leaf.as_str()) || declares(bare) {
-        return true;
-    }
-    if proto == "Codec" {
-        let has = |m: &str| env.functions.contains_key(&sym(&format!("{}.{}", leaf, m)))
-            || env.functions.contains_key(&sym(&format!("{}.{}", bare, m)));
-        return has("encode") && has("decode");
-    }
-    false
-}
-/// The Codec derive serializes a field by structural recursion over String/Int/Float/Bool/Option/List/Named — it has NO Map or Set arm, so a `Map[K,V]` / `Set[T]` field silently falls through to the `Value`-as-String fallback: invalid Rust natively (E0614/E0308) and wrong/silent on wasm (#655). Detect such a container anywhere in the field type (under List/Option/Result/Tuple/anon-record). A `Map`/`Set` reached only through a NAMED type is that type's own concern (its `: Codec` is checked by the leaf rule), so we stop at `Ty::Named`.
-fn codec_unsupported_container(ty: &Ty) -> Option<&'static str> {
-    use almide_lang::types::TypeConstructorId as TC;
-    match ty {
-        Ty::Applied(TC::Map, _) => Some("Map"),
-        Ty::Applied(TC::Set, _) => Some("Set"),
-        Ty::Applied(_, args) => args.iter().find_map(|a| codec_unsupported_container(a)),
-        Ty::Tuple(elems) => elems.iter().find_map(|e| codec_unsupported_container(e)),
-        Ty::Record { fields } | Ty::OpenRecord { fields } =>
-            fields.iter().find_map(|(_, t)| codec_unsupported_container(t)),
-        _ => None,
-    }
-}
-/// A type that derives a field-recursive protocol (Codec/Ord/Hash) requires every field type to ALSO satisfy it — otherwise the derive emits a call to a non-existent `Field.encode` (Codec) or a Rust `#[derive(Ord/Hash)]` over a field whose Rust type lacks the impl, both of which the checker previously accepted and codegen then rejected as "invalid Rust" (#611). This validates the requirement structurally, at the checker, independent of target.
-fn validate_derive_field_support(env: &TypeEnv, diagnostics: &mut Vec<Diagnostic>) {
-    let pairs: Vec<(Sym, Vec<Sym>)> = env.type_protocols.iter()
-        .map(|(ty, protos)| (*ty, protos.iter().copied().collect()))
-        .collect();
-    let mut reported: std::collections::HashSet<(Sym, Sym, Sym)> = std::collections::HashSet::new();
-    for (type_name, protocols) in &pairs {
-        let Some(ty) = env.types.get(type_name) else { continue };
-        let slots = type_field_slots(ty);
-        if slots.is_empty() { continue; }
-        for proto in protocols {
-            let p = proto.as_str();
-            if !FIELD_RECURSIVE_PROTOCOLS.contains(&p) { continue; }
-            for (field_name, field_ty) in &slots {
-                let site = DeriveSite { type_name: *type_name, proto: *proto };
-                validate_derive_field(env, diagnostics, &mut reported, site, field_name, field_ty);
-            }
-        }
-    }
-}
-/// Per-field derive-support check for a single (type, protocol, field) triple: the Codec-unsupported-container check, plus recursively requiring every leaf nominal in the field's type to derive the same protocol. `reported` dedupes diagnostics across (type, protocol, leaf) triples seen by earlier fields/protocols/types in the caller's loop nest. Verbatim text move out of [`validate_derive_field_support`].
-/// The (type, protocol) pair a field-recursive derive check is running for.
-///
-/// Both are `Sym`, so as adjacent positional parameters they could be
-/// transposed silently — the check would then report the type as the protocol.
-#[derive(Copy, Clone)]
-struct DeriveSite {
-    type_name: Sym,
-    proto: Sym,
-}
-
-fn validate_derive_field(
-    env: &TypeEnv, diagnostics: &mut Vec<Diagnostic>,
-    reported: &mut std::collections::HashSet<(Sym, Sym, Sym)>,
-    site: DeriveSite, field_name: &str, field_ty: &Ty,
-) {
-    let DeriveSite { type_name, proto } = site;
-    // The protocol arrived twice — as a `Sym` and as its own `&str` — so the
-    // two could be passed from different protocols at the call site. One name,
-    // derived here.
-    let p = proto.as_str();
-    // #655: the Codec derive has no Map/Set arm — reject such a field here rather than emitting invalid Rust / silent-wrong wasm. Same E023 family (a field that cannot satisfy Codec).
-    if p == "Codec" {
-        if let Some(container) = codec_unsupported_container(field_ty) {
-            if reported.insert((type_name, proto, sym(container))) {
-                diagnostics.push(err(
-                    format!("type '{}' derives 'Codec' but field '{}' has a '{}' type, which the Codec derive cannot encode",
-                        type_name, field_name, container),
-                    format!("The Codec derive serializes a {} as a String, which is wrong. Use a List[(K, V)] field (or List[T] for a Set), or implement encode/decode manually.",
-                        container),
-                    format!("type {} : Codec", type_name),
-                ).with_code("E023"));
-            }
-        }
-    }
-    let mut leaves = Vec::new();
-    collect_leaf_nominals(field_ty, &mut leaves);
-    for leaf in leaves {
-        if leaf == type_name { continue; }          // self-reference is fine
-        if !env.types.contains_key(&leaf) { continue; } // not a user nominal → native support
-        if leaf_satisfies(env, leaf, p) { continue; }
-        if !reported.insert((type_name, proto, leaf)) { continue; }
-        diagnostics.push(err(
-            format!("type '{}' derives '{}' but field '{}' has type '{}', which does not derive '{}'",
-                type_name, p, field_name, leaf, p),
-            format!("Add `: {}` to the declaration of type '{}' (every field of a `: {}` type must itself be `{}`)",
-                p, leaf, p, p),
-            format!("type {} : {}", type_name, p),
-        ).with_code("E023"));
-    }
-}
-/// Validate that types declaring `: ProtocolName` have all required convention methods, AND that each present method's signature actually matches the protocol's declared signature (arity, parameter types, return type) — `Self` substituted for the declaring type. Called after all declarations are registered so all `Type.method` functions are available. Signature checking is skipped for generic types (`contains_typevar`): `Self` would need to carry the type's own type arguments (e.g. `Self` for `Container[X]` must resolve to `Container[X]`, not bare `Container`), and nothing currently threads a user type's declared generic parameters back in here. Presence is still required either way.
-pub fn validate_protocol_impls(env: &TypeEnv, diagnostics: &mut Vec<Diagnostic>) {
-    let type_protocols: Vec<(Sym, Vec<Sym>)> = env.type_protocols.iter()
-        .map(|(ty, protos)| (*ty, protos.iter().copied().collect()))
-        .collect();
-
-    for (type_name, protocol_names) in &type_protocols {
-        let is_generic = env.types.get(type_name).is_some_and(|t| t.contains_typevar());
-        let type_ty = Ty::Named(*type_name, vec![]);
-
-        for proto_name in protocol_names {
-            let proto_def = match env.protocols.get(proto_name) {
-                Some(p) => p.clone(),
-                None => continue,
-            };
-
-            for method_sig in &proto_def.methods {
-                let target = ImplTarget { name: *type_name, is_generic, ty: &type_ty };
-                validate_protocol_method_impl(env, diagnostics, &target, *proto_name, method_sig);
-            }
-        }
-    }
-}
-/// Validate a single protocol method's implementation on `type_name`: presence (missing-method E023 unless it's a builtin-derived protocol), then — for non-generic types — arity, parameter types, and return type against the protocol's declared signature (`Self` substituted for `type_ty`). Verbatim text move out of [`validate_protocol_impls`].
-/// The type a protocol implementation is being validated on.
-///
-/// `is_generic` gates the signature comparison: a generic type's method
-/// signature carries unsubstituted type vars, so comparing it against the
-/// protocol's declaration would report spurious mismatches.
-struct ImplTarget<'a> {
-    name: Sym,
-    is_generic: bool,
-    ty: &'a Ty,
-}
-
-fn validate_protocol_method_impl(
-    env: &TypeEnv, diagnostics: &mut Vec<Diagnostic>,
-    target: &ImplTarget<'_>, proto_name: Sym,
-    method_sig: &crate::types::ProtocolMethodSig,
-) {
-    let ImplTarget { name: type_name, is_generic, ty: type_ty } = *target;
-    let fn_key = format!("{}.{}", type_name, method_sig.name);
-    let Some(sig) = env.functions.get(&sym(&fn_key)) else {
-        let is_builtin = matches!(proto_name.as_str(),
-            "Eq" | "Repr" | "Ord" | "Hash" | "Codec" | "Encode" | "Decode"
-            | "Numeric");
-        if !is_builtin {
-            diagnostics.push(err(
-                format!("type '{}' declares protocol '{}' but missing method '{}'",
-                    type_name, proto_name, method_sig.name),
-                format!("Add: fn {}.{}({}) -> {}",
-                    type_name, method_sig.name,
-                    method_sig.params.iter()
-                        .map(|(n, t)| {
-                            let display_ty = if *t == Ty::TypeVar(sym("Self")) {
-                                type_name.to_string()
-                            } else {
-                                t.display()
-                            };
-                            format!("{}: {}", n, display_ty)
-                        })
-                        .collect::<Vec<_>>()
-                        .join(", "),
-                    {
-                        let ret = &method_sig.ret;
-                        if *ret == Ty::TypeVar(sym("Self")) {
-                            type_name.to_string()
-                        } else {
-                            ret.display()
-                        }
-                    }),
-                format!("type {} : {}", type_name, proto_name),
-            ));
-        }
-        return;
-    };
-    if is_generic {
-        return;
-    }
-
-    let expected_params: Vec<Ty> = method_sig.params.iter()
-        .map(|(_, ty)| env.resolve_named(&substitute_self(ty, type_ty)))
-        .collect();
-    let expected_ret = env.resolve_named(&substitute_self(&method_sig.ret, type_ty));
-    let actual_params: Vec<Ty> = sig.params.iter().map(|(_, t)| env.resolve_named(t)).collect();
-    let actual_ret = env.resolve_named(&sig.ret);
-
-    if actual_params.len() != expected_params.len() {
-        diagnostics.push(err(
-            format!("method '{}' on type '{}' has {} parameter(s), expected {} to satisfy protocol '{}'",
-                method_sig.name, type_name, actual_params.len(), expected_params.len(), proto_name),
-            format!("Protocol '{}' defines: fn {}({})", proto_name, method_sig.name,
-                method_sig.params.iter().map(|(n, t)| {
-                    format!("{}: {}", n, substitute_self(t, type_ty).display())
-                }).collect::<Vec<_>>().join(", ")),
-            format!("fn {}.{}", type_name, method_sig.name),
-        ));
-        return;
-    }
-    for (i, (actual_ty, expected_ty)) in actual_params.iter().zip(expected_params.iter()).enumerate() {
-        let matches = strip_module_qualifier(actual_ty) == strip_module_qualifier(expected_ty);
-        if !matches && *expected_ty != Ty::Unknown && *actual_ty != Ty::Unknown {
-            let param_name = &sig.params[i].0;
-            diagnostics.push(err(
-                format!("method '{}.{}' parameter '{}' has type '{}', expected '{}' to satisfy protocol '{}'",
-                    type_name, method_sig.name, param_name, actual_ty.display(), expected_ty.display(), proto_name),
-                format!("Change type to '{}'", expected_ty.display()),
-                format!("fn {}.{}", type_name, method_sig.name),
-            ));
-        }
-    }
-    let ret_matches = strip_module_qualifier(&actual_ret) == strip_module_qualifier(&expected_ret);
-    if !ret_matches && expected_ret != Ty::Unknown && actual_ret != Ty::Unknown {
-        diagnostics.push(err(
-            format!("method '{}.{}' returns '{}', expected '{}' to satisfy protocol '{}'",
-                type_name, method_sig.name, actual_ret.display(), expected_ret.display(), proto_name),
-            format!("Change return type to '{}'", expected_ret.display()),
-            format!("fn {}.{}", type_name, method_sig.name),
-        ));
-    }
-}
 /// A borrowed view of the `type` declaration being registered.
 pub struct TypeDeclToRegister<'a> {
     pub name: &'a str,
@@ -845,3 +586,5 @@ fn register_decl_top_let(env: &mut TypeEnv, decl: &ast::Decl, prefix: Option<&st
     let did = env.def_table.alloc(sym(pkg), sym(mod_path), sym(name), almide_ir::DefKind::TopLet, rt);
     env.def_map.insert(sym(&key), did);
 }
+
+include!("registration_validate.rs");
